requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
jinja2==3.1.4
openpyxl==3.1.5
reportlab==4.2.2
apscheduler==3.10.4
aiofiles==24.1.0
python-dotenv==1.0.1
